Move properties from AppConfig into endpoint service
Most of te new code in EndpointDiscoveryService.java has been moved
from AppConfig.java.
