+++ src/main/java/health/ere/ps/service/connector/endpoints/EndpointDiscoveryService.java
@@ -4,6 +4,8 @@ import health.ere.ps.config.AppConfig;
 import health.ere.ps.exception.common.security.SecretsManagerException;
 import health.ere.ps.service.common.security.SecretsManagerService;
 import health.ere.ps.ssl.SSLUtilities;
+import org.apache.commons.lang3.StringUtils;
+import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
@@ -20,6 +22,7 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.Optional;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -27,18 +30,36 @@ import java.util.logging.Logger;
 public class EndpointDiscoveryService {
     private static final Logger LOG = Logger.getLogger(EndpointDiscoveryService.class.getName());
 
-    private String signatureServiceEndpoint;
+    /**
+     * Certificate to authenticate at the connector.
+     */
+    @ConfigProperty(name = "connector.cert.auth.store.file")
+    Optional<String> connectorTlsCertTrustStore;
 
-    private String eventServiceEndpoint;
+    /**
+     * Password of the certificate to authenticate at the connector.
+     * The default value is a empty sting, so that the password must not be set.
+     */
+    @ConfigProperty(name = "connector.cert.auth.store.file.password", defaultValue = "")
+    String connectorTlsCertTustStorePwd;
 
-    private String certificateServiceEndpoint;
+    @ConfigProperty(name = "auth-signature-service.endpoint.address", defaultValue = "")
+    String authSignatureServiceEndpointAddress;
 
-    private String authSignatureServiceEndpoint;
+    @ConfigProperty(name = "card-service.endpoint.address", defaultValue = "")
+    String cardServiceEndpointAddress;
 
-    private String cardServiceEndpoint;
+    @ConfigProperty(name = "certificate-service.endpoint.address", defaultValue = "")
+    String certificateServiceEndpointAddress;
 
-    @Inject
-    AppConfig appConfig;
+    @ConfigProperty(name = "event-service.endpoint.address", defaultValue = "")
+    String eventServiceEndpointAddress;
+
+    @ConfigProperty(name = "connector.base-uri")
+    String connectorBaseUri;
+
+    @ConfigProperty(name = "connector.verify-hostname", defaultValue = "true")
+    String connectorVerifyHostname;
 
     @Inject
     SecretsManagerService secretsManagerService;
@@ -48,34 +69,24 @@ public class EndpointDiscoveryService {
         // code copied from IdpClient.java
 
         // having an ssl context does not interfere with non-ssl connections
-        SSLContext sslContext = appConfig.getConnectorTlsCertTrustStore().isPresent() ? secretsManagerService.createSSLContext(appConfig.getConnectorTlsCertTrustStore().get(),
-                appConfig.getConnectorTlsCertTustStorePwd(),
+        SSLContext sslContext = connectorTlsCertTrustStore.isPresent() ? secretsManagerService.createSSLContext(connectorTlsCertTrustStore.get(),
+                connectorTlsCertTustStorePwd,
                 SecretsManagerService.SslContextType.TLS,
                 SecretsManagerService.KeyStoreType.PKCS12) : secretsManagerService.createAcceptAllSSLContext();
 
         ClientBuilder clientBuilder = ClientBuilder.newBuilder()
                 .sslContext(sslContext);
 
-        if (!appConfig.isConnectorVerifyHostname()) {
+        if ("false".equals(connectorVerifyHostname)) {
             // disable hostname verification
             clientBuilder = clientBuilder.hostnameVerifier(new SSLUtilities.FakeHostnameVerifier());
         }
 
         Invocation invocation = clientBuilder.build()
-                .target(appConfig.getConnectorBaseUri())
+                .target(connectorBaseUri)
                 .path("/connector.sds")
                 .request("application/xml")
                 .buildGet();
-
-        String signatureServiceEndpoint = "";
-
-        String authSignatureServiceEndpoint = "";
-
-        String cardServiceEndpoint = "";
-
-        String eventServiceEndpoint = "";
-
-        String certificateServiceEndpoint = "";
 
         try (InputStream inputStream = invocation.invoke(InputStream.class)) {
             Document document = DocumentBuilderFactory.newDefaultInstance()
@@ -98,24 +109,20 @@ public class EndpointDiscoveryService {
                 }
 
                 switch (node.getAttributes().getNamedItem("Name").getTextContent()) {
-                    case "SignatureService": {
-                        signatureServiceEndpoint = getEndpoint(node);
-                        break;
-                    }
                     case "AuthSignatureService": {
-                        authSignatureServiceEndpoint = getEndpoint(node);
+                        authSignatureServiceEndpointAddress = getEndpoint(node);
                         break;
                     }
                     case "CardService": {
-                        cardServiceEndpoint = getEndpoint(node);
+                        cardServiceEndpointAddress = getEndpoint(node);
                         break;
                     }
                     case "EventService": {
-                        eventServiceEndpoint = getEndpoint(node);
+                        eventServiceEndpointAddress = getEndpoint(node);
                         break;
                     }
                     case "CertificateService": {
-                        certificateServiceEndpoint = getEndpoint(node);
+                        certificateServiceEndpointAddress = getEndpoint(node);
                         break;
                     }
                 }
@@ -124,32 +131,30 @@ public class EndpointDiscoveryService {
         } catch (SAXException | IllegalArgumentException e) {
             LOG.log(Level.SEVERE, "Could not parse connector.sds", e);
         }
-
-        this.signatureServiceEndpoint = signatureServiceEndpoint;
-        this.authSignatureServiceEndpoint = authSignatureServiceEndpoint;
-        this.cardServiceEndpoint = cardServiceEndpoint;
-        this.eventServiceEndpoint = eventServiceEndpoint;
-        this.certificateServiceEndpoint = certificateServiceEndpoint;
     }
 
-    public String getSignatureServiceEndpoint() {
-        return signatureServiceEndpoint;
+    public String getAuthSignatureServiceEndpointAddress() {
+        return authSignatureServiceEndpointAddress;
     }
 
-    public String getEventServiceEndpoint() {
-        return eventServiceEndpoint;
+    public String getCardServiceEndpointAddress() {
+        return cardServiceEndpointAddress;
     }
 
-    public String getCertificateServiceEndpoint() {
-        return certificateServiceEndpoint;
+    public String getCertificateServiceEndpointAddress() {
+        return certificateServiceEndpointAddress;
     }
 
-    public String getAuthSignatureServiceEndpoint() {
-        return authSignatureServiceEndpoint;
+    public String getEventServiceEndpointAddress() {
+        return eventServiceEndpointAddress;
     }
 
-    public String getCardServiceEndpoint() {
-        return cardServiceEndpoint;
+    public Optional<String> getConnectorTlsCertTrustStore() {
+        return connectorTlsCertTrustStore;
+    }
+
+    public String getConnectorTlsCertTrustStorePwd() {
+        return StringUtils.defaultString(connectorTlsCertTustStorePwd).trim();
     }
 
     private String getEndpoint(Node serviceNode) {
@@ -173,7 +178,7 @@ public class EndpointDiscoveryService {
 
             String location = endpointList.item(0).getAttributes().getNamedItem("Location").getTextContent();
 
-            if (location.startsWith(appConfig.getConnectorBaseUri())) {
+            if (location.startsWith(connectorBaseUri)) {
                 return location;
             }
         }
